Remove duplicated environment modules
