+++ lib/sanbase/queries/queries.ex
@@ -11,6 +11,7 @@ defmodule Sanbase.Queries do
   alias Sanbase.Queries.QueryExecution
   alias Sanbase.Queries.DashboardQueryMapping
   alias Sanbase.Accounts.User
+  alias Sanbase.Clickhouse.Query.Environment
 
   import Sanbase.Utils.ErrorHandling, only: [changeset_errors_string: 1]
 
@@ -70,7 +71,7 @@ defmodule Sanbase.Queries do
   def run_query(%Query{} = query, user, query_metadata, opts \\ []) do
     query_metadata = Map.put_new(query_metadata, :sanbase_user_id, user.id)
 
-    with {:ok, environment} <- Sanbase.Queries.Environment.new(query, user),
+    with {:ok, environment} <- Environment.new(query, user),
          {:ok, result} <- Sanbase.Queries.Executor.run(query, query_metadata, environment) do
       maybe_store_execution_data_async(result, user.id, opts)
